efficient version and base model is downloaded to dataset dir
