+++ ultralytics/cfg/default.yaml
@@ -7,7 +7,7 @@ task: detect # (str) YOLO task, i.e. detect, segment, classify, pose, obb
 mode: val # (str) YOLO mode, i.e. train, val, predict, export, track, benchmark
 datasets_dir:  #/Users/yamtawachi/tensorleap/ #added by tensorleap
 # Train settings -------------------------------------------------------------------------------------------------------
-model: # (str, optional) path to model file, i.e. yolov8n.pt, yolov8n.yaml
+model: yolo11s.pt  # set to the core of your.h5 model    # (str, optional) path to model file, i.e. yolov8n.pt, yolov8n.yaml
 data: coco.yaml # (str, optional) path to data file, i.e. coco8.yaml
 epochs: 100 # (int) number of epochs to train for
 time: # (float, optional) number of hours to train for, overrides epochs if supplied
